updating order_items

+++ src/main/java/com/wissen/instantvendingmachine/repository/OrderItemsRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.wissen.instantvendingmachine.entity.OrderItemsEntity;
 
 @Repository
-public interface OrderItemsRepository extends JpaRepository<OrderItemsEntity, Integer> {
-
+public interface OrderItemsRepository extends JpaRepository<OrderItemsEntity, Long> {
 }
